home-manager: fix runit services for 1st gen

+++ nixpkgs/lib/runit.nix
@@ -135,15 +135,17 @@ in
           done
 
           # Cleanup old services; logs are cleaned first
-          oldSvcPath=$oldGenPath/home-files/.local/var/service
-          newSvcPath=$newGenPath/home-files/.local/var/service
-          for svc in "$oldSvcPath"/*/log "$oldSvcPath"/*; do
-            svcName=''${svc##$oldSvcPath/}
-            if [[ ! -e "$newSvcPath/$svcName" ]]; then
-              $DRY_RUN_CMD rm $VERBOSE_ARG "${dstDir}/$svcName"/supervise
-              $DRY_RUN_CMD rmdir -p --ignore-fail-on-non-empty $VERBOSE_ARG "${dstDir}/$svcName"
-            fi
-          done
+          if [[ -n "$oldGenPath" ]]; then
+            oldSvcPath=$oldGenPath/home-files/.local/var/service
+            newSvcPath=$newGenPath/home-files/.local/var/service
+            for svc in "$oldSvcPath"/*/log "$oldSvcPath"/*; do
+              svcName=''${svc##$oldSvcPath/}
+              if [[ -d "$svc" ]] && [[ ! -e "$newSvcPath/$svcName" ]]; then
+                $DRY_RUN_CMD rm $VERBOSE_ARG "${dstDir}/$svcName"/supervise
+                $DRY_RUN_CMD rmdir -p --ignore-fail-on-non-empty $VERBOSE_ARG "${dstDir}/$svcName"
+              fi
+            done
+          fi
         }
 
         setupRunitServices
